🎨 Rename module: go-std/src -> go-std/app
This change sets the default executable name to "app".
We may also drop the `-o app` in the Dockerfile.
