+++ frameworks/Go/go-std/src/storage/mysql.go
@@ -7,7 +7,7 @@ import (
 	"math/rand"
 	"sort"
 
-	"go-std/src/templates"
+	"go-std/app/templates"
 
 	_ "github.com/go-sql-driver/mysql" // postgresql import
 )
