+++ frameworks/Go/go-std/src/storage/pq.go
@@ -7,7 +7,7 @@ package storage
 // 	"math/rand"
 // 	"sort"
 
-// 	"go-std/src/templates"
+// 	"go-std/app/templates"
 
 // 	_ "github.com/lib/pq" // postgresql import
 // )
